暂时不用通讯模块 +圆形进度条

+++ Voystock.Communication/Indicator.cs
@@ -15,6 +15,6 @@ namespace Voystock.Communication
 		public string Name { get; set; }
 
 		[DataMember]
-		public List<int> Parameters { get; set; }
+		public List<float> Parameters { get; set; }
 	}
 }
